Configure service container for tests
This works by first instantiating a postgres container and having GitHub
Actions wait for the postgres container to be made available.

Then we need to make sure to pass in environment variables for the step
that executes the command, so we do that by setting step level
environment variables and then citing those commands in the docker run
command so that we can pass those, ultimately, into the settings.py for
the tests to be able to create a test database to run the tests on.

The other little "trick" here is the `--network host` which tells the
docker container that's running in the step to use the host network
which is where we're running postgres itself. Since we're utilizing the
host network, then to the docker container that executes the tests, it
appears as though we're simply just connecting to localhost:5432 to
connect to postgres.