Setup traefik for reverse proxy
In order to keep it simple to serve both the frontend and the backend
from the same domain, I've configured traefik to run in the local
cluster to act as the reverse proxy which will route between the
different applications.

By leveraging the match rules in Traefik, we can say that `/api/.*` will
route to the Django application. Everything else will route to the
frontend.