Use a local docker registry for faster image pulls
ctlptl is a tool that's provided by the Tilt engineering team that does
two major things. Firstly, it wraps creation of building local clusters.
We used to do this with kind and have switched to using ctlptl. This on
its own is not very interesting.

The interesting part comes from the fact that ctlptl will ALSO create a
local docker registry for you and link it with the local kind cluster
that it creates. Having a local docker registry means that entire layers
can be cached at once instead of what was happening in the past, where
every time a docker image was built, the image in its entirety needed to
be loaded into the local kind cluster.

By running a local docker registry, only layers that have changed need
to be pushed again. We need to slightly update our helm charts in order
to tell Tilt to build, push, and pull to the appropriate registry.