Make tilt down an explicit recipe
Previously, every single time you exited out of tilt, we'd shutdown the
resources for you. This meant that the next time you wanted to do a
`make up`, that there would have to be some time spent in your
applications and databases all starting up again. We did this
automatically so that if tilt itself wasn't running, the cluster/dev
environment was not using up any resources on your computer.

We've opted to make the down step explicit which will remove all running
applications from your cluster. This way, even if you exit out of tilt,
you won't have have to wait for everything to re-initialize the next
time you want to run a `make up`. And if you really need it, you can run
`make down` to free up some resources.