Don't full rebuild when dependencies change
Previously, we used the `fall_back_on` directive which would cause the
entire Docker image to fully rebuild if there was a detected change in
the files that listed dependencies. This worked but in the case of a
dependency change, the full rebuild was extremely slow, relative to the
other hot reloading of the Tilt based setup.

Instead, we'll conditionally trigger a re-install based on the files
that changed. This way, the entire image does not need to be
re-installed in the moment, but dependency updates will still be
propagated directly to the environment of the application that's being
run in a pod.